ttodo-be_merge1/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jwt==1.3.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
